+++ apps/back-office/README.md
@@ -1,0 +1,1 @@
+# back-office
