+++ apps/back-office/biome.json
@@ -1,0 +1,3 @@
+{
+	"extends": ["../../biome.json"]
+}
