+++ apps/back-office/src/app/page.tsx
@@ -1,0 +1,9 @@
+import { ProfileForm } from "~/components/form-example";
+
+export default function Page() {
+	return (
+		<div className="py-32 space-y-7">
+			<ProfileForm />
+		</div>
+	);
+}
