+++ apps/back-office/src/components/form-example.tsx
@@ -1,0 +1,409 @@
+// @see https://ui.shadcn.com/docs/components/form
+"use client";
+
+import Link from "next/link";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { format } from "date-fns";
+import * as z from "zod";
+import { CalendarIcon, CheckIcon, ChevronsUpDownIcon } from "lucide-react";
+
+import { cn } from "~/libs/utils";
+
+import { Button } from "~/components/ui/button";
+import {
+	Form,
+	FormControl,
+	FormDescription,
+	FormField,
+	FormItem,
+	FormLabel,
+	FormMessage,
+} from "~/components/ui/form";
+import { Input } from "~/components/ui/input";
+import { Textarea } from "~/components/ui/textarea";
+import { Checkbox } from "~/components/ui/checkbox";
+import { RadioGroup, RadioGroupItem } from "~/components/ui/radio-group";
+import {
+	Select,
+	SelectContent,
+	SelectItem,
+	SelectTrigger,
+	SelectValue,
+} from "~/components/ui/select";
+import { Switch } from "~/components/ui/switch";
+import {
+	Command,
+	CommandEmpty,
+	CommandGroup,
+	CommandInput,
+	CommandItem,
+} from "~/components/ui/command";
+import {
+	Popover,
+	PopoverContent,
+	PopoverTrigger,
+} from "~/components/ui/popover";
+import { Calendar } from "~/components/ui/calender";
+import { toast } from "sonner";
+
+const languages = [
+	{ label: "English", value: "en" },
+	{ label: "French", value: "fr" },
+	{ label: "German", value: "de" },
+	{ label: "Spanish", value: "es" },
+	{ label: "Portuguese", value: "pt" },
+	{ label: "Russian", value: "ru" },
+	{ label: "Japanese", value: "ja" },
+	{ label: "Korean", value: "ko" },
+	{ label: "Chinese", value: "zh" },
+] as const;
+
+const formSchema = z.object({
+	username: z.string().min(2, {
+		message: "Username must be at least 2 characters.",
+	}),
+	bio: z
+		.string()
+		.min(10, {
+			message: "Bio must be at least 10 characters.",
+		})
+		.max(160, {
+			message: "Bio must not be longer than 30 characters.",
+		}),
+	mobile: z.boolean().default(false).optional(),
+	type: z.enum(["all", "mentions", "none"], {
+		required_error: "You need to select a notification type.",
+	}),
+	email: z
+		.string({
+			required_error: "Please select an email to display.",
+		})
+		.email(),
+	marketing_emails: z.boolean().default(false).optional(),
+	security_emails: z.boolean(),
+	language: z.string({
+		required_error: "Please select a language.",
+	}),
+	dob: z.date({
+		required_error: "A date of birth is required.",
+	}),
+});
+
+export function ProfileForm() {
+	// 1. Define your form.
+	const form = useForm<z.infer<typeof formSchema>>({
+		resolver: zodResolver(formSchema),
+		defaultValues: {
+			username: "",
+			bio: "",
+			security_emails: true,
+		},
+	});
+
+	// 2. Define a submit handler.
+	async function onSubmit(values: z.infer<typeof formSchema>) {
+		// Do something with the form values.
+		// ✅ This will be type-safe and validated.
+		console.log(values);
+
+		toast.promise(
+			new Promise((resolve) => setTimeout(() => resolve(values), 1000)),
+			{
+				loading: "Loading...",
+				success: (data: any) => {
+					return `${data.username} has been added, check out the console.`;
+				},
+				error: "Error",
+			},
+		);
+	}
+
+	return (
+		<Form {...form}>
+			<form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
+				<FormField
+					control={form.control}
+					name="username"
+					render={({ field }) => (
+						<FormItem>
+							<FormLabel>Username</FormLabel>
+							<FormControl>
+								<Input placeholder="shadcn" {...field} />
+							</FormControl>
+							<FormDescription>
+								This is your public display name.
+							</FormDescription>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<FormField
+					control={form.control}
+					name="bio"
+					render={({ field }) => (
+						<FormItem>
+							<FormLabel>Bio</FormLabel>
+							<FormControl>
+								<Textarea
+									placeholder="Tell us a little bit about yourself"
+									className="resize-none"
+									{...field}
+								/>
+							</FormControl>
+							<FormDescription>
+								You can <span>@mention</span> other users and organizations.
+							</FormDescription>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<FormField
+					control={form.control}
+					name="mobile"
+					render={({ field }) => (
+						<FormItem className="flex flex-row items-start space-x-3 space-y-0 rounded-md border p-4 shadow">
+							<FormControl>
+								<Checkbox
+									checked={field.value}
+									onCheckedChange={field.onChange}
+								/>
+							</FormControl>
+							<div className="space-y-1 leading-none">
+								<FormLabel>
+									Use different settings for my mobile devices
+								</FormLabel>
+								<FormDescription>
+									You can manage your mobile notifications in the{" "}
+									<Link href="/">mobile settings</Link> page.
+								</FormDescription>
+							</div>
+						</FormItem>
+					)}
+				/>
+				<FormField
+					control={form.control}
+					name="type"
+					render={({ field }) => (
+						<FormItem className="space-y-3">
+							<FormLabel>Notify me about...</FormLabel>
+							<FormControl>
+								<RadioGroup
+									onValueChange={field.onChange}
+									defaultValue={field.value}
+									className="flex flex-col space-y-1"
+								>
+									<FormItem className="flex items-center space-x-3 space-y-0">
+										<FormControl>
+											<RadioGroupItem value="all" />
+										</FormControl>
+										<FormLabel className="font-normal">
+											All new messages
+										</FormLabel>
+									</FormItem>
+									<FormItem className="flex items-center space-x-3 space-y-0">
+										<FormControl>
+											<RadioGroupItem value="mentions" />
+										</FormControl>
+										<FormLabel className="font-normal">
+											Direct messages and mentions
+										</FormLabel>
+									</FormItem>
+									<FormItem className="flex items-center space-x-3 space-y-0">
+										<FormControl>
+											<RadioGroupItem value="none" />
+										</FormControl>
+										<FormLabel className="font-normal">Nothing</FormLabel>
+									</FormItem>
+								</RadioGroup>
+							</FormControl>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<FormField
+					control={form.control}
+					name="email"
+					render={({ field }) => (
+						<FormItem>
+							<FormLabel>Email</FormLabel>
+							<Select onValueChange={field.onChange} defaultValue={field.value}>
+								<FormControl>
+									<SelectTrigger>
+										<SelectValue placeholder="Select a verified email to display" />
+									</SelectTrigger>
+								</FormControl>
+								<SelectContent>
+									<SelectItem value="m@example.com">m@example.com</SelectItem>
+									<SelectItem value="m@google.com">m@google.com</SelectItem>
+									<SelectItem value="m@support.com">m@support.com</SelectItem>
+								</SelectContent>
+							</Select>
+							<FormDescription>
+								You can manage email addresses in your{" "}
+								<Link href="/examples/forms">email settings</Link>.
+							</FormDescription>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<div>
+					<h3 className="mb-4 text-lg font-medium">Email Notifications</h3>
+					<div className="space-y-4">
+						<FormField
+							control={form.control}
+							name="marketing_emails"
+							render={({ field }) => (
+								<FormItem className="flex flex-row items-center justify-between rounded-lg border p-3 shadow-sm">
+									<div className="space-y-0.5">
+										<FormLabel>Marketing emails</FormLabel>
+										<FormDescription>
+											Receive emails about new products, features, and more.
+										</FormDescription>
+									</div>
+									<FormControl>
+										<Switch
+											checked={field.value}
+											onCheckedChange={field.onChange}
+										/>
+									</FormControl>
+								</FormItem>
+							)}
+						/>
+						<FormField
+							control={form.control}
+							name="security_emails"
+							render={({ field }) => (
+								<FormItem className="flex flex-row items-center justify-between rounded-lg border p-3 shadow-sm">
+									<div className="space-y-0.5">
+										<FormLabel>Security emails</FormLabel>
+										<FormDescription>
+											Receive emails about your account security.
+										</FormDescription>
+									</div>
+									<FormControl>
+										<Switch
+											checked={field.value}
+											onCheckedChange={field.onChange}
+											disabled
+											aria-readonly
+										/>
+									</FormControl>
+								</FormItem>
+							)}
+						/>
+					</div>
+				</div>
+				<FormField
+					control={form.control}
+					name="language"
+					render={({ field }) => (
+						<FormItem className="flex flex-col">
+							<FormLabel>Language</FormLabel>
+							<Popover>
+								<PopoverTrigger asChild>
+									<FormControl>
+										<Button
+											variant="outline"
+											role="combobox"
+											className={cn(
+												"w-[200px] justify-between",
+												!field.value && "text-muted-foreground",
+											)}
+										>
+											{field.value
+												? languages.find(
+														(language) => language.value === field.value,
+												  )?.label
+												: "Select language"}
+											<ChevronsUpDownIcon className="ml-2 h-4 w-4 shrink-0 opacity-50" />
+										</Button>
+									</FormControl>
+								</PopoverTrigger>
+								<PopoverContent className="w-[200px] p-0">
+									<Command>
+										<CommandInput
+											placeholder="Search framework..."
+											className="h-9"
+										/>
+										<CommandEmpty>No framework found.</CommandEmpty>
+										<CommandGroup>
+											{languages.map((language) => (
+												<CommandItem
+													value={language.label}
+													key={language.value}
+													onSelect={() => {
+														form.setValue("language", language.value);
+													}}
+												>
+													{language.label}
+													<CheckIcon
+														className={cn(
+															"ml-auto h-4 w-4",
+															language.value === field.value
+																? "opacity-100"
+																: "opacity-0",
+														)}
+													/>
+												</CommandItem>
+											))}
+										</CommandGroup>
+									</Command>
+								</PopoverContent>
+							</Popover>
+							<FormDescription>
+								This is the language that will be used in the dashboard.
+							</FormDescription>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<FormField
+					control={form.control}
+					name="dob"
+					render={({ field }) => (
+						<FormItem className="flex flex-col">
+							<FormLabel>Date of birth</FormLabel>
+							<Popover>
+								<PopoverTrigger asChild>
+									<FormControl>
+										<Button
+											variant={"outline"}
+											className={cn(
+												"w-[240px] pl-3 text-left font-normal",
+												!field.value && "text-muted-foreground",
+											)}
+										>
+											{field.value ? (
+												format(field.value, "PPP")
+											) : (
+												<span>Pick a date</span>
+											)}
+											<CalendarIcon className="ml-auto h-4 w-4 opacity-50" />
+										</Button>
+									</FormControl>
+								</PopoverTrigger>
+								<PopoverContent className="w-auto p-0" align="start">
+									<Calendar
+										mode="single"
+										selected={field.value}
+										onSelect={field.onChange}
+										disabled={(date) =>
+											date > new Date() || date < new Date("1900-01-01")
+										}
+										initialFocus
+									/>
+								</PopoverContent>
+							</Popover>
+							<FormDescription>
+								Your date of birth is used to calculate your age.
+							</FormDescription>
+							<FormMessage />
+						</FormItem>
+					)}
+				/>
+				<Button type="submit">Submit</Button>
+			</form>
+		</Form>
+	);
+}
