+++ biome.json
@@ -1,4 +1,19 @@
 {
   "$schema": "https://biomejs.dev/schemas/1.5.2/schema.json",
-  "extends": ["./formatter.json", "./linter.json"]
+  "organizeImports": { "enabled": true },
+  "formatter": {
+    "enabled": true,
+    "indentWidth": 2,
+    "formatWithErrors": false,
+    "indentStyle": "tab",
+    "lineWidth": 80,
+    "lineEnding": "lf",
+    "ignore": ["node_modules/**/*", "dist/**/*", ".next/**/*"]
+  },
+  "linter": {
+    "enabled": true,
+    "rules": {
+      "recommended": true
+    }
+  }
 }
