--- formatter.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://biomejs.dev/schemas/1.5.2/schema.json",
-  "formatter": {
-    "enabled": true,
-    "indentWidth": 2,
-    "formatWithErrors": false,
-    "indentStyle": "tab",
-    "lineWidth": 80,
-    "lineEnding": "lf",
-    "ignore": []
-  }
-}
