--- linter.json
@@ -1,9 +1,0 @@
-{
-  "$schema": "https://biomejs.dev/schemas/1.5.2/schema.json",
-  "linter": {
-    "enabled": true,
-    "rules": {
-      "recommended": true
-    }
-  }
-}
