--- packages/back-office/src/app/layout.tsx
@@ -1,14 +1,0 @@
-import '~/styles/globals.css';
-import { cn } from '~/libs/utils';
-import { Toaster } from '~/components/ui/sonner';
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang='ko'>
-      <body className={cn('container min-h-screen bg-background font-sans antialiased')}>
-        {children}
-        <Toaster />
-      </body>
-    </html>
-  );
-}
