--- packages/back-office/src/app/page.tsx
@@ -1,9 +1,0 @@
-import { ProfileForm } from '~/components/form-example';
-
-export default function Page() {
-  return (
-    <div className='py-32 space-y-7'>
-      <ProfileForm></ProfileForm>
-    </div>
-  );
-}
